Updates based on review comments (v0.2)
This is the design revision that was ordered for the first prototype build.

The pinout of the cell connector and the board-to-board connector changed
compared to v0.1, hence it got a minor revision update.

The control board is now a 4-layer board to improve the grounding of the
PCB. Various other suggested improvements from the design review are also
included in this update.

Fixes #11, #12, #13, #17, #18, #19