ci: use origin/main for worktree in detached mode
Previously, a workaround was implemented by pinning to a fixed version because the fix in https://github.com/INTI-CMNB/KiBot/commit/d66188014c67307d9820ad9a72239d624488c446 wasn't included in the LTS tag. With the release of v2_k8_1_6_5, the fix is now integrated, requiring the update of `main` to `origin/main` in `diff.kibot.yaml`.